requirements.txt
Django==4.1.5
djangorestframework==3.14.0
Pillow==9.4.0
django-galleryfield
requests
httpsig